p4i integration Integrate 249135 from 4.2 to main: Fixes:    compile for Windows RevBy:    thomas Details:
[git-p4: depot-paths = "//depot/qt/main/": change = 249138]

+++ src/corelib/tools/qdatetime.cpp
@@ -4570,7 +4570,7 @@ QDateTimeParser::State QDateTimeParser::checkIntermediate(const QDateTime &dt, c
     for (int i=0; i<sectionNodes.size(); ++i) {
         const SectionNode &sn = sectionNodes.at(i);
         QString t = sectionText(s, i, sn.pos).toLower();
-        if (t.contains(space) || (t.size() < sectionMaxSize(i) && fieldInfo(i) & (FixedWidth|Numeric) != Numeric)) {
+        if (t.contains(space) || (t.size() < sectionMaxSize(i) && (((int)fieldInfo(i) & (FixedWidth|Numeric)) != Numeric))) {
             found = true;
             switch (sn.type) {
             case AmPmSection:
